Add category management form and update product handling
- Added a new form `frmCategory` in `SDCafeOffice.Views` for managing product categories.
- Introduced methods in `frmCategory` for loading, saving, inserting, and deleting categories, interacting with `DataAccessPOS`.
- Updated `DataAccessPOS.cs` with new methods for handling categories and updated existing methods to include new fields.
- Added new properties `TaxCode` and `CategoryId` to `POS_ProductModel` and updated related methods.
- Modified `POS_ProductTypeModel` to change `SortOrder` type and add `ForeColor` and `BackColor` properties.
- Added a new file `POS_CategoryModel.cs` to define the category model.
- Updated various forms (`frmProduct`, `frmLogOn`, `frmMain`, etc.) to include new UI elements and logic for category and tax code handling.
- Adjusted data grid views and event handlers to accommodate new category and tax code fields.
- Enhanced UI elements in several forms, including setting `ControlBox` to false and adding icons.

+++ Office/SDCafeOffice/Views/frmProdType.cs
@@ -17,6 +17,8 @@ namespace SDCafeOffice.Views
     {
         private string strTypeId;
         List<POS_ProductTypeModel> ptypes = new List<POS_ProductTypeModel>();
+        private Color m_ButtonBackColor;
+        private Color m_ButtonForeColor;
         public frmProdType()
         {
             InitializeComponent();
@@ -47,10 +49,37 @@ namespace SDCafeOffice.Views
             {
                 txt_TypeID.Text = ptypes[0].Id.ToString();
                 txt_TypeName.Text = ptypes[0].TypeName;
+                txt_SortOrder.Text = ptypes[0].SortOrder.ToString();
                 check_Liquor.Checked = ptypes[0].IsLiquor;
                 check_Restaurant.Checked = ptypes[0].IsRestaurant;
                 check_Donation.Checked = ptypes[0].IsBatchDonation;
                 check_Discount.Checked = ptypes[0].IsBatchDiscount;
+
+                if (string.IsNullOrEmpty(ptypes[0].ForeColor))
+                {
+                    picForeColor.BackColor = Color.Black;
+                    m_ButtonForeColor = Color.Black;
+                }
+                else
+                {
+                    // convert prods[0].ForeColor rgb string to color
+                    picForeColor.BackColor = Color.FromArgb(int.Parse(ptypes[0].ForeColor));
+                    m_ButtonForeColor = Color.FromArgb(int.Parse(ptypes[0].ForeColor));
+                }
+                if (string.IsNullOrEmpty(ptypes[0].BackColor))
+                {
+                    picBackColor.BackColor = Color.White;
+                    m_ButtonBackColor = Color.White;
+                }
+                else
+                {
+                    // convert prods[0].BackColor rgb string to color
+                    picBackColor.BackColor = Color.FromArgb(int.Parse(ptypes[0].BackColor));
+                    m_ButtonBackColor = Color.FromArgb(int.Parse(ptypes[0].BackColor));
+                }
+                btButtonColor.BackColor = m_ButtonBackColor;
+                btButtonColor.ForeColor = m_ButtonForeColor;
+                btButtonColor.Text = ptypes[0].TypeName;
             }
         }
 
@@ -73,16 +102,36 @@ namespace SDCafeOffice.Views
             //if (String.IsNullOrEmpty(txt_ConfigName.Text)) txt_ConfigName.Text = "";
             //if (String.IsNullOrEmpty(txt_ConfigValue.Text)) txt_ConfigValue.Text = "";
             //if (String.IsNullOrEmpty(txt_ConfigDesc.Text)) txt_ConfigDesc.Text = "";
-
+            int iId = 0;
+            int iSortOrder = 0;
+            try
+            {
+                iId = int.Parse(txt_TypeID.Text);
+            }
+            catch (Exception ex)
+            {
+                iId = 0;
+            }
+            try
+            {
+                iSortOrder = int.Parse(txt_SortOrder.Text);
+            }
+            catch (Exception ex)
+            {
+                iSortOrder = 0;
+            }
             ptypes.Clear();
             ptypes.Add(new POS_ProductTypeModel()
             {
-                Id = int.Parse(txt_TypeID.Text),
+                Id = iId,
                 TypeName = txt_TypeName.Text,
                 IsLiquor = check_Liquor.Checked,
+                SortOrder = iSortOrder,
                 IsRestaurant = check_Restaurant.Checked,
                 IsBatchDonation = check_Donation.Checked,
-                IsBatchDiscount = check_Discount.Checked
+                IsBatchDiscount = check_Discount.Checked,
+                ForeColor = m_ButtonForeColor.ToArgb().ToString(),
+                BackColor = m_ButtonBackColor.ToArgb().ToString()
             });
             int iCnt = dbPOS.Update_ProductType(ptypes[0]);
             txtMessage.Text = "Type successfully Updated : " + txt_TypeName.Text;
@@ -96,15 +145,36 @@ namespace SDCafeOffice.Views
             //if (String.IsNullOrEmpty(txt_ConfigName.Text)) txt_ConfigName.Text = "";
             //if (String.IsNullOrEmpty(txt_ConfigValue.Text)) txt_ConfigValue.Text = "";
             //if (String.IsNullOrEmpty(txt_ConfigDesc.Text)) txt_ConfigDesc.Text = "";
-
+            int iId = 0;
+            int iSortOrder = 0;
+            try
+            {
+                iId = int.Parse(txt_TypeID.Text);
+            }
+            catch (Exception ex)
+            {
+                iId = 0;
+            }
+            try
+            {
+                iSortOrder = int.Parse(txt_SortOrder.Text);
+            }
+            catch (Exception ex)
+            {
+                iSortOrder = 0;
+            }
             ptypes.Clear();
             ptypes.Add(new POS_ProductTypeModel()
             {
+                Id = iId,
                 TypeName = txt_TypeName.Text,
                 IsLiquor = check_Liquor.Checked,
+                SortOrder = iSortOrder,
                 IsRestaurant = check_Restaurant.Checked,
                 IsBatchDonation = check_Donation.Checked,
-                IsBatchDiscount = check_Discount.Checked
+                IsBatchDiscount = check_Discount.Checked,
+                ForeColor = m_ButtonForeColor.ToArgb().ToString(),
+                BackColor = m_ButtonBackColor.ToArgb().ToString()
             });
             int iCnt = dbPOS.Insert_ProductType(ptypes[0]);
             txtMessage.Text = "Type successfully Added : " + txt_TypeName.Text;
@@ -132,13 +202,48 @@ namespace SDCafeOffice.Views
 
         private void bt_Delete_Click(object sender, EventArgs e)
         {
+            DataAccessPOS dbPOS = new DataAccessPOS();
             int iSelectedProdTypeId = Convert.ToInt32(txt_TypeID.Text);
+            int iAssociatedProds =   dbPOS.Get_All_Products_By_ProdType(iSelectedProdTypeId).Count;
+            if (iAssociatedProds > 0)
+            {
+                // Cannot delete Product Type which has associated products
+                txtMessage.Text = "Cannot delete Product Type which has associated products ! " + iAssociatedProds.ToString();
+                return;
+            }
             if (iSelectedProdTypeId > 0)
             {
-                DataAccessPOS dbPOS = new DataAccessPOS();
                 dbPOS.Delete_ProductType_By_Id(iSelectedProdTypeId);
                 bt_Exit.PerformClick();
             }
+        }
+
+        private void picForeColor_Click(object sender, EventArgs e)
+        {
+            ColorDialog clrDialog = new ColorDialog();
+
+            //show the colour dialog and check that user clicked ok
+            if (clrDialog.ShowDialog() == DialogResult.OK)
+            {
+                //save the colour that the user chose
+                m_ButtonForeColor = clrDialog.Color;
+                picForeColor.BackColor = m_ButtonForeColor;
+            }
+            btButtonColor.ForeColor = m_ButtonForeColor;
+        }
+
+        private void picBackColor_Click(object sender, EventArgs e)
+        {
+            ColorDialog clrDialog = new ColorDialog();
+
+            //show the colour dialog and check that user clicked ok
+            if (clrDialog.ShowDialog() == DialogResult.OK)
+            {
+                //save the colour that the user chose
+                m_ButtonBackColor = clrDialog.Color;
+                picBackColor.BackColor = m_ButtonBackColor;
+            }
+            btButtonColor.BackColor = m_ButtonBackColor;
         }
     }
 }
